+++ lib/go/thrift/header_transport_test.go
@@ -21,6 +21,7 @@ package thrift
 
 import (
 	"context"
+	"io"
 	"io/ioutil"
 	"testing"
 )
@@ -73,9 +74,20 @@ func TestTHeaderHeadersReadWrite(t *testing.T) {
 	}
 
 	// Read
+
+	// Make sure multiple calls to ReadFrame is fine.
+	if err := reader.ReadFrame(); err != nil {
+		t.Errorf("reader.ReadFrame returned error: %v", err)
+	}
+	if err := reader.ReadFrame(); err != nil {
+		t.Errorf("reader.ReadFrame returned error: %v", err)
+	}
 	read, err := ioutil.ReadAll(reader)
 	if err != nil {
 		t.Errorf("Read returned error: %v", err)
+	}
+	if err := reader.ReadFrame(); err != nil && err != io.EOF {
+		t.Errorf("reader.ReadFrame returned error: %v", err)
 	}
 	if string(read) != payload1+payload2 {
 		t.Errorf(
@@ -104,5 +116,15 @@ func TestTHeaderHeadersReadWrite(t *testing.T) {
 			"reader.GetReadHeaders() expected size 2, actual content: %+v",
 			headers,
 		)
+	}
+}
+
+func TestTHeaderTransportNoDoubleWrapping(t *testing.T) {
+	trans := NewTMemoryBuffer()
+	orig := NewTHeaderTransport(trans)
+	wrapped := NewTHeaderTransport(orig)
+
+	if wrapped != orig {
+		t.Errorf("NewTHeaderTransport double wrapped THeaderTransport")
 	}
 }